server/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
websockets==12.0
python-jose==3.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
Pillow==10.1.0
gunicorn==21.2.0